backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
python-dotenv>=1.0,<2.0
httpx>=0.28,<1.0
python-multipart>=0.0.20,<1.0
pillow>=11.0,<12.0
opencv-python>=4.10,<5.0
torch>=2.3,<3.0
ultralytics>=8.3,<9.0
supabase>=2.28,<3.0
apscheduler>=3.11,<4.0
resend>=2.29,<3.0
google-generativeai>=0.8,<1.0
openwakeword>=0.6,<1.0
pyaudio>=0.2.14,<1.0
numpy>=1.26,<3.0
